+++ src/cli/main.pl
@@ -68,13 +68,20 @@ opt_spec(serve,'terminusdb serve OPTIONS',
            shortflags([h]),
            longflags([help]),
            default(false),
-           help('print help for `serve` command')],
+           help('Print help for `serve` command')],
           [opt(interactive),
            type(boolean),
            shortflags([i]),
            longflags([interactive]),
            default(false),
-           help('run server in interactive mode')]]).
+           help('Run server in interactive mode')],
+          [opt(memory),
+           type(atom),
+           shortflags([m]),
+           longflags([memory]),
+           meta(password),
+           default('_'),
+           help('Run server in-memory, without a persistent store. Takes a password as an optional argument. The in-memory store will be initialized with an admin account with the given password. If absent, the admin account will have \'root\' as a password.')]]).
 opt_spec(list,'terminusdb list OPTIONS',
          'List databases.',
          [[opt(help),
@@ -669,13 +676,15 @@ run_command(test,_Positional,Opts) :-
     ->  halt(0)
     ;   halt(1)).
 run_command(serve,_Positional,Opts) :-
-    (   option(interactive(true),Opts)
+    (   option(interactive(true), Opts)
     ->  terminus_server([serve|Opts], false),
         prolog
     ;   terminus_server([serve|Opts], true)).
 run_command(list,Databases,_Opts) :-
     super_user_authority(Auth),
-    list_databases(system_descriptor{}, Auth, Databases, Database_Objects),
+    (   Databases = []
+    ->  list_databases(system_descriptor{}, Auth, Database_Objects)
+    ;   list_existing_databases(Databases, Database_Objects)),
     pretty_print_databases(Database_Objects).
 run_command(optimize,Databases,_Opts) :-
     super_user_authority(Auth),